Revert "Fixing minor issues causing failing tests"
This reverts commit ebe816283f2db4553f6727b4875c0b18e1550ab7.

+++ src/Umbraco.Cms.Persistence.EFCore.SqlServer/Migrations/UmbracoDbContextModelSnapshot.cs
@@ -302,7 +302,7 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("nvarchar(100)")
                         .HasColumnName("languageCultureName");
 
-                    b.Property<int?>("FallbackLanguageId")
+                    b.Property<short?>("FallbackLanguageId")
                         .HasColumnType("smallint")
                         .HasColumnName("fallbackLanguageId");
